+++ Sources/Tokenizer/Layout/ConstraintPriority.swift
@@ -6,9 +6,9 @@ public enum ConstraintPriority {
     case high
     case medium
     case low
-    case custom(Double)
+    case custom(Float)
 
-    public var numeric: Double {
+    public var numeric: Float {
         switch self {
         case .required:
             return 1000.0
