+++ Sources/Tokenizer/Styles/SupportedPropertyType.swift
@@ -77,19 +77,19 @@ public enum SupportedPropertyType {
         case .collapseAxis:
             return ViewCollapseAxis(rawValue: text).map(SupportedPropertyValue.collapseAxis)
         case .rect:
-            let parts = text.components(separatedBy: ",").flatMap(Double.init)
+            let parts = text.components(separatedBy: ",").flatMap(Float.init)
             guard parts.count == 4 else { return nil }
             return .rect(Rect(x: parts[0], y: parts[1], width: parts[2], height: parts[3]))
         case .point:
-            let parts = text.components(separatedBy: ",").flatMap(Double.init)
+            let parts = text.components(separatedBy: ",").flatMap(Float.init)
             guard parts.count == 2 else { return nil }
             return .point(Point(x: parts[0], y: parts[1]))
         case .size:
-            let parts = text.components(separatedBy: ",").flatMap(Double.init)
+            let parts = text.components(separatedBy: ",").flatMap(Float.init)
             guard parts.count == 2 else { return nil }
             return .size(Size(width: parts[0], height: parts[1]))
         case .edgeInsets:
-            let parts = text.components(separatedBy: ",").flatMap(Double.init)
+            let parts = text.components(separatedBy: ",").flatMap(Float.init)
             guard parts.count == 4 || parts.count == 2 else { return nil }
             if parts.count == 4 {
                 return .edgeInsets(EdgeInsets(top: parts[0], left: parts[1], bottom: parts[2], right: parts[3]))
