+++ Sources/Tokenizer/Util/Float+CGFloat.swift
@@ -1,5 +1,5 @@
 //
-//  Double+CGFloat.swift
+//  Float+CGFloat.swift
 //  Reactant
 //
 //  Created by Matouš Hýbl on 23/04/2017.
@@ -8,7 +8,7 @@
 import CoreGraphics
 import Foundation
 
-extension Double {
+extension Float {
     var cgFloat: CGFloat {
         return CGFloat(self)
     }
